requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
paho-mqtt==2.1.0
PyYAML==6.0.3
requests==2.32.5
urllib3==2.5.0
